Split server actions into their own files
