+++ apps/web/app/actions/auth.ts
@@ -1,0 +1,28 @@
+"use server";
+
+import { encodedRedirect } from "@/utils";
+import { createClientForServer } from "@/utils/supabase/server";
+import { redirect } from "next/navigation";
+
+export const signInAction = async (formData: FormData) => {
+  const email = formData.get("email") as string;
+  const password = formData.get("password") as string;
+  const supabase = await createClientForServer();
+
+  const { error } = await supabase.auth.signInWithPassword({
+    email,
+    password,
+  });
+
+  if (error) {
+    return encodedRedirect("error", "/sign-in", error.message);
+  }
+
+  return redirect("/");
+};
+
+export const signOutAction = async () => {
+  const supabase = await createClientForServer();
+  await supabase.auth.signOut();
+  return redirect("/sign-in");
+};
