+++ apps/web/app/actions/index.ts
@@ -1,0 +1,4 @@
+// Re-export all actions for backward compatibility
+export * from './auth';
+export * from './items';
+export * from './search';
