+++ apps/web/app/actions/items.ts
@@ -3,30 +3,7 @@
 import { encodedRedirect } from "@/utils";
 import { createClientForServer } from "@/utils/supabase/server";
 import { redirect } from "next/navigation";
-import { Item, ItemInsert, ItemUpdate } from "@/types";
-
-export const signInAction = async (formData: FormData) => {
-  const email = formData.get("email") as string;
-  const password = formData.get("password") as string;
-  const supabase = await createClientForServer();
-
-  const { error } = await supabase.auth.signInWithPassword({
-    email,
-    password,
-  });
-
-  if (error) {
-    return encodedRedirect("error", "/sign-in", error.message);
-  }
-
-  return redirect("/");
-};
-
-export const signOutAction = async () => {
-  const supabase = await createClientForServer();
-  await supabase.auth.signOut();
-  return redirect("/sign-in");
-};
+import { ItemInsert, ItemUpdate } from "@/types";
 
 export const createItemAction = async (formData: FormData) => {
   const itemType = formData.get("itemtype")?.toString();
@@ -120,50 +97,3 @@ export const updateItemAction = async (formData: FormData) => {
 
   return redirect(`/year/${belongsToYear}`);
 };
-
-export type SearchState = {
-  query: string;
-  results: Item[];
-  initial: boolean;
-};
-export async function searchItems(
-  prevState: SearchState,
-  formData: FormData,
-): Promise<SearchState> {
-  const query = formData.get("query") as string;
-
-  if (!query) return { query: "", results: [], initial: false };
-
-  // Trim and validate query length
-  const trimmedQuery = query.trim();
-  if (trimmedQuery.length < 2) {
-    return { query, results: [], initial: false };
-  }
-
-  const supabase = await createClientForServer();
-
-  // Escape special characters for ILIKE pattern matching
-  const escapedQuery = trimmedQuery.replace(/[%_]/g, "\\$&");
-
-  const { data, error } = await supabase
-    .from("items")
-    .select(
-      "id, title, author, director, itemtype, season, published_year, belongs_to_year, redo",
-    ) // Select only needed columns
-    .or(
-      `title.ilike.%${escapedQuery}%,author.ilike.%${escapedQuery}%,director.ilike.%${escapedQuery}%`,
-    )
-    .order("created_at", { ascending: false })
-    .limit(100); // Limit results to prevent excessive data transfer
-
-  if (error) {
-    console.error("Search error:", error);
-    return { query, results: [], initial: false };
-  }
-
-  return {
-    query,
-    results: data as Item[],
-    initial: false,
-  };
-}
