+++ apps/web/components/forms/create-item-form.tsx
@@ -2,7 +2,7 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Checkbox } from "@/components/ui/checkbox";
-import { createItemAction } from "@/app/actions";
+import { createItemAction } from "@/app/actions/items";
 import { SubmitButton } from "./submit-button";
 
 export default function CreateItemForm() {
