+++ apps/web/components/lists/items-list.tsx
@@ -4,55 +4,24 @@ import { CategoryList } from "@/components/lists/category-list";
 export default async function ItemsList({ year }: { year: number }) {
   const supabase = await createClientForServer();
 
-  const categories = [
-    { title: "Books", type: "Book" },
-    { title: "Movies", type: "Movie" },
-    { title: "TV Shows", type: "Show" },
-  ];
-
-  // Fetch items for each category in parallel with database-level filtering
+  // Fetch all items for the year and user data
   const [
-    { data: books, error: booksError },
-    { data: movies, error: moviesError },
-    { data: shows, error: showsError },
+    { data: items, error },
     {
       data: { user },
     },
   ] = await Promise.all([
     supabase
       .from("items")
-      .select(
-        "id, title, author, itemtype, published_year, belongs_to_year, redo, created_at",
-      )
+      .select("*")
       .eq("belongs_to_year", year)
-      .eq("itemtype", "Book")
-      .order("created_at", { ascending: true }),
-    supabase
-      .from("items")
-      .select(
-        "id, title, director, itemtype, published_year, belongs_to_year, redo, created_at",
-      )
-      .eq("belongs_to_year", year)
-      .eq("itemtype", "Movie")
-      .order("created_at", { ascending: true }),
-    supabase
-      .from("items")
-      .select(
-        "id, title, season, itemtype, published_year, belongs_to_year, redo, created_at",
-      )
-      .eq("belongs_to_year", year)
-      .eq("itemtype", "Show")
       .order("created_at", { ascending: true }),
     supabase.auth.getUser(),
   ]);
 
   // Check for any errors
-  if (booksError || moviesError || showsError) {
-    console.error("Error fetching items:", {
-      booksError,
-      moviesError,
-      showsError,
-    });
+  if (error) {
+    console.error("Error fetching items:", error);
     return (
       <div role="alert" className="text-center p-8">
         <p className="text-red-600">Unable to load your items right now.</p>
@@ -63,11 +32,24 @@ export default async function ItemsList({ year }: { year: number }) {
     );
   }
 
-  // Map categories to their respective data
+  // Group items by category
+  const allItems = items || [];
   const categoryData = [
-    { title: "Books", type: "Book", items: books || [] },
-    { title: "Movies", type: "Movie", items: movies || [] },
-    { title: "TV Shows", type: "Show", items: shows || [] },
+    { 
+      title: "Books", 
+      type: "Book" as const, 
+      items: allItems.filter(item => item.itemtype === "Book") 
+    },
+    { 
+      title: "Movies", 
+      type: "Movie" as const, 
+      items: allItems.filter(item => item.itemtype === "Movie") 
+    },
+    { 
+      title: "TV Shows", 
+      type: "Show" as const, 
+      items: allItems.filter(item => item.itemtype === "Show") 
+    },
   ];
 
   return (
